requirements.txt
# Core dependencies for Fed Minutes Parser
PyMuPDF>=1.23.0
pandas>=2.0.0
numpy>=1.24.0
pyyaml>=6.0

# Phase 2: Knowledge Base dependencies
sentence-transformers>=2.2.0
chromadb>=0.4.0
scikit-learn>=1.3.0

# Phase 3: AI Analysis dependencies
openai>=1.0.0
anthropic>=0.25.0
python-dotenv>=1.0.0

# For testing
pytest>=7.0.0

# For notebooks (optional)
jupyter>=1.0.0
matplotlib>=3.5.0
seaborn>=0.12.0